do_read wip

+++ include/thread_memory.h
@@ -17,7 +17,7 @@ public:
     }
 
     void add_to_store_buffer(ModelAction *action);
-	uint8_t get_last_write(ModelAction* act);
+	uint64_t get_last_write(ModelAction* act);
     bool pop_from_store_buffer();
     void empty_store_buffer();
 };
